+++ models/agent/src/info/mod.rs
@@ -1,0 +1,12 @@
+mod agent;
+mod datastore;
+mod shard;
+
+pub use self::agent::AgentInfo;
+pub use self::agent::AgentVersion;
+pub use self::datastore::DatastoreInfo;
+pub use self::shard::CommitOffset;
+pub use self::shard::CommitUnit;
+pub use self::shard::Shard;
+pub use self::shard::ShardRole;
+pub use self::shard::Shards;