+++ models/agent/src/lib.rs
@@ -1,17 +1,2 @@
-extern crate serde;
-extern crate serde_derive;
-#[cfg(test)]
-extern crate serde_json;
-
-mod agent;
-mod datastore;
-mod shard;
-
-pub use self::agent::AgentInfo;
-pub use self::agent::AgentVersion;
-pub use self::datastore::DatastoreInfo;
-pub use self::shard::CommitOffset;
-pub use self::shard::CommitUnit;
-pub use self::shard::Shard;
-pub use self::shard::ShardRole;
-pub use self::shard::Shards;
+pub mod actions;
+pub mod info;