+++ models/agent/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Actions models.
+
+### Changed
+- **BREAKING**: Re-arrange information models into `info` module.
 
 ## 0.3.0 - 2019-06-16
 ### Added